thermal_cells - Added generic AppController class to control flow
This is a generic app controller to sit between the PApplet (of Processing) that has the setup() and draw() methods, and the data controller for this particular application.

This allows for the continual call of draw() for UI controls, but perhaps not advancing the data model because the app is effectively Paused.