p5js - Ecosystem - Introduce basic erosion
This is achieved by:
- Establishing concept of 'neighbors' in the CellGrid
- For each cell, determine its 'lowestNeighbor'
- Starting from the highest cell, transfer some elevation from that
cell to its 'lowestNeighbor'

- This is performed if a new 'erosionRate' is set.
... (values between 0 - 0.3 are best)

Additionally it has the concept of 'cumulativeInflux' which is an
approximation of the amount of rainfall (assuming uniform rain) that
flows through that cell. From that, we can perform additional
erosion with the idea that it is washed away in a river.

Currently if the erosion runs for too many cycles ... noise ends
up geting introduced.