p5js/water-cycle - Derive the contents of a WeatherCell from Terrain
Use the pixels array of the Terrain, and newly added Color utils (in P5JsUtils object) to determine if a particular cell contains air / water / soil.

This is a simple check of the center pixel of the cell.

(Originally, I intended to have concepts of air-water or air-soil, or water-soil boundaries within the cells, but backed off of that complexity).

Limitation:
- There are some cells that do not match the air / water / soil colors ... I suspect this is due to some anti-aliasing, perhaps from the default "smooth" behavior in P5JS.