processing/starfied - First pass at Coding Challenge #1
This is a first pass at the Coding Train's Challenge #1 - Starfield simulator
https://www.youtube.com/watch?v=17WoOqgXsRM

This solution uses simple 2D space, and fakes the motion by moving pixesl out along their radial line from the center of the screen.

Shiffman's solution in the YouTube video makes use of 3D modeling, and projection onto a 2D canvas, and thus allows for better depth-of-field effect of stars closer to the viewer.