p5js/mitosis - Add concept of division axis (spindle axis)
* During the Metaphase, the Centrosomes move to opposite poles. They then extend spindles to meet in the center.

This commit establishes the virtual line connecting the Centrosomses as the 'axisOfSpindles'

This commit also changes how the first FurrowPoint is selected. Rather than a random one, it now finds the one that is closest to a 90-degree (PI/2) rotation from the 'heading' (p5.js notation of a Vector's rotation) of the 'axisOfSpindles'.