processing/hockey-01 - Add preliminary constraint check for Rink
* When moving the Puck, ensure that the new location is valid.

Right now, it just does an optimistic check for one of the 3 large (simple) rectangle spaces, and if its in there, everything is OK ... e.g. don't do any bouncing.

NOTE:
I'm intending to try to determine what the "bouncing" effect is, and thus passing in the previous position and velocity ... but not using them just yet.