processing/hockey-01 - Fix yet another bug in corner collision
This was an interesting consideration.

The net effect of this is that I'm using a slightly smaller radius when determing the (x,y) of the Puck when it hits the boards.

The curious part (I think) is that the radius of the effective (trimmed) corner is the original corner's radius, lessened by the radius of the object bouncing off it.

This makes sense. The (x,y) of the Puck when it hits the wall will be that distance from the corner's center.

A similar approach should be taken when hitting straight walls as well.