p5js/forest-fires - Introduce basic propagation of fire
On each tick of the clock, a given cell can ignite if a sufficient number of its neighbors are burning.

Uses an 'influence matrix' to weight the influence of the 8 (grid) neighoring cells, with emphasis given on cardinal direction neighbors.

Also updates common JS model: CellGrid to have a convenience method of `cellNeighbordsOfIdx()`   ... mainly because I mistakenly thought `neighborsOfIdx()` already did that.

This results in a very digital propagation fo fire ... and currently requires two Lightning strikes to overcome the 0.1 'fireRiskThreshold' to ignite a cell with fuel.