p5js/forest-fires - Refactor UI to specific UserInterface class
This puts a class between the p5.js events (keyPressed / mouseClicked) and the System.

I intend to introduce a UI control panel, and tool menu bar, and thus want to have an object manage concepts like the currently selected Tool, and not have this float in a global 'currentTool' variable.