p5js/forest-fires - Add UI tools to create a Fire Break
Allows the user to create a firebreak to prevent the spread of fire.

User selects this tool by pressing 'F', and then clicks on the map where they want to create the firebreak.

This effectively removes the fuel from that cell, and transitions it to a SOIL terrain type; thus it won't ignite and act like a wall (until wind comes into the system and allows fire to jump breaks).

More Info: https://en.wikipedia.org/wiki/Firebreak