p5js/forest-fires - Add UI tool to 'knockdown' the fire
This enables the user to zero-out the fire intensity at a given point, and  thus transition the cell into the SMOLDERING state.

That cell can re-ignite on the next tick() ... and this is often the case, so currently requires combination with the firebreak to be effective against large fires.

Glossary: https://www.fs.fed.us/nwacfire/home/terminology.html#K