p5js/forest-fires - Only render cells that have changed type
Each cell now keeps track of whether or not it needs to be rendered.

NOTE: this pushes some rendering logic into the data model ... what if a 'renderer' wanted to visualize the change in a property other than the 'terrainType'?

This does result in a significant reduction in CPU usage. On my computer, it drops from 100% (of one core) to ~25% ...

Might be premature optimization, but going to stick with this for now.
(Also noting that another approach could be to use the 'pixels' array, since the `rect()` function for p5.js seems to lag with high number of cells, ~10,000)