p5js/forest-fires - Add pixel rendering but don't use (for now)
(I don't normally like adding code that isn't used ... but this is a "sketchbook" afterall, and as noted it may be of use when there are large number of cells being rendered in a given draw() cycle.

This adds an alternate mode, using p5.js' direct pixel array manipulation: https://p5js.org/reference/#/p5/pixels

It comes at a cost of having to call loadPixels() and updatePixels(), which moves a lot of memory around, and thus a bit slower when not much is being updated.

* Render All Views ... Pixel mode is faster than rect() per cell
* Render views as needed ... rect() is faster (when number of cells updated is small).

That implies some tipping point where the # of cells to redraw, (or perhaps the percent of the pixels being modified) warrants using the pixel rendering.

That ability to switch between the modes (and perhaps do so dynamically per cycle) warranted leaving it as a mode rather than create the pixel renderer as a subclass.

Additional note: The pre-calculation of the RGB values of the color-scheme was warranted (rather than red(), green(), blue() of direct lookup) and borne out of JS profiling that showed p5.js color functions using a high percentage of CPU time.