p5js/forest-fires - Add 'tool' to ignite fire at mouse location
This willl more likely be used for testing ... it will be useful to start fires in particular areas of the map.

Pressing 'L' sets the tool mode to LIGHTNING ... clicking on a cell where there is some fuel (e.g. FOLLIAGE) will ignite it.

Not to be confused with future implementation of controlled burns, to remove the fuel in a controlled manner ahead of a fire.