p5js/forest-fires-02 - Add display of UI control panel
This adds a control panel on the right side of the sketch, including a group of buttons.

For now, this using the default buttons and UISet I've previously established ...

Some issues / challenges there:
- The UISet assumed global coordinates; Originally, I intended to use local coordinates;
- Buttons own their styling, but I can subclass here.

NOTE: I need to arrive at consistent local/global coordinate system for nested UI; global seems to be simpler to manage in the long run; and I haven't had extensive nested UIs (yet).