p5js/forest-fires-02 - Add concept of limited Resources
Previously, the user could create perform the 'FIRE_BREAK' action without any limit.

Now, they start off with 10, and slowly overtime, get more actions.

If they are out of the action for a given resource, they can not use that resource.

This is intended to force the user to be decisive about where to deploy particular resources.