p5js/forest-fires-02 - Allow basic editing of terrain
Intent here is to a first step towards a level editor ...

So the user can take a randomly generated map, edit it, and eventually save it out. For now, they can just edit it in the current session.

Users press D, and then a number 0-7 (corresponding to a System.TERRAIN_* type)

  TERRAIN_SOIL: 0
  TERRAIN_WATER: 1
  TERRAIN_FOLLIAGE: 2
  TERRAIN_BURNING: 3
  TERRAIN_ENGULFED: 4
  TERRAIN_SMOLDERING: 5
  TERRAIN_BURNT: 6
  TERRAIN_PARTIAL_BURN: 7