p5js/tree-rings - Add dat.gui, bound to System.settings
Adds dat.gui for config control.

See: https://github.com/dataarts/dat.gui

NOTE: This approach of binding to the 'system.settings' object is new to me ... previously I would create a distinct object, defined with the app.js with defaults ... this approach allows the existing P5JsSettings to be observed (from the query string) and layers on the dat.gui binding.