p5js/forest-01 - Adjust placement of dat.gui to bottom right
To avoid having the dat.gui <div> overlay the main nav of the site, this places it in the bottom right of window.

NOTE: Had to reimplement the Hide/Show toggle, because it appears that when disabling the autoplacement, it does not observe the 'H' key to hide/show. (And calling: dat.GUI.toggleHide(); won't hide it either).

Examples from doc of custom placement:
http://workshop.chromeexperiments.com/examples/gui/#8--Custom-Placement