p5js/forest-01 - Refactor out dat.gui placement to P5JsSettings
This moves the custom placement of thedat.gui config panel to the common P5JsSettings object.

NOTE: For now, I'm going with raw JS to create/modify the elements, rather than jQuery ... this is intended to reach a broader audience (meaning there isn't a rabbit hole of learning), and limit dependencies ... though I may revisit this. Similar with the inline HTML / CSS attributes, rather than move to a .css file (and thus another reference) this keeps the logic contained.