p5js/forest-fires-02 - Fix bug after randomization
When 'Randomizing' after loading a scenario, the scenario would not appear to load.

What was happening was that the System.randomize() was modifying the ScenarioManager's copy of the settings when it was creating a new random seed.

This fixes the issue by using a simple (shallow copy) of the settings object. This is okay because we won't ever (ever?) have a nested objects in our settings.

Lifted from:
https://medium.com/better-programming/3-ways-to-clone-objects-in-javascript-f752d148054d