p5js/forest-fires-02 - Add a new simple intro scenario
This adds a simple intro scenario which has a lightning strike in a FOLIAGE area that is connected to 3 other wooded areas via narrow breaks in the SOIL cover.

This should provide a good intro to players to know that they can leverage the existing terrain to control the spread of fires.

NOTE: It also uses the recent change to fire risk which only requires 1 lightning strike to start the fire. An implication of this is that the player can use the KNOCK_DOWN suppression tactic to reduce the intensity of the first strike and put out the fire before it spreads.  (Currently the game speed is a bit to quick to easily do this).