p5js/forest-fires-02 - Refactor UI to use DOM elements
Replaces the Canvas drawn buttons with DOM <button> elements.

This reduces the burden on repainting the buttons to the canvas.

An alternate approach would be to keep track of whether or not the buttons needed re-rendered.

But an additional benefit is that the buttons are rendered using a consisten approach, and can be styled consistently once I get to that stage.