p5js/forest-fires-02 - Adjust setting of initial fuel amount
This fixes a bug with the draw terrain tool, `SHIFT + D`, then `0-7`; where drawing WATER (2) on top of FOLIAGE would not clear out the fuel amount, and thus the cell could catch on fire.

Also, it simplifies the inital generation of the map, and not require setting the fuel amount specifically for FOLIAGE.

I kept this within the `Systme.init()` because I do see it as configuration whereby the initila fuel amounts per terrain type could vary from scenario to scenario.

Finally, it does improve the loading of a Terrain image file a bit, by giving the cells that are on fire some fuel to burn ... but it does not set their `fire_intensity`.