p5js/watershed - Generate a first ridge line
Establishes a basic concept of a 'Ridge'

Currently, it just has an x,y location
Eventually it will have an elevation, as well as a direct connection up and down hill.

NOTE: The generation currently places 'Ridges' beyond the 'bounds'. There needs to be a method by which it detects that it has left the displayable area, effectively "hike" back up the ridgeline, and branch out.

(Though I do wonder if there is some benefit to generating surrounding area to know what is beyond the bounds of the displayable area).