p5js/watershed - Establish elevation of ridge (points) and color gradient
Uses Perlin noise to set an elevation of each point.

A new 'ColorGradient' class is established to map a value to a color; with a default mapping being that of  general terrain from -1 (deep sea) to 0 (sea level, coast, to 1 (snow-capped mountain top).

During rendering, a small circle is placed that the Ridge's coordinates corresponding to the elevation mapped onto the ColorGradient.