p5js/nautical-flags - Add rendering of (one) Special NATO Flag: SQ
This adds the rendering of one of the special NATO flags, to lay some foundation work for adding others.

It introduces the concept of `CTRL` + 2-LETTERS to request a special flag be drawn.

It currently just has support for the SQ combination, for the Squandron flag.

MINOR: There was a necessary patch the 'supportedFlags' regex, becaus now we are listening to keys like 'SHIFT' which matched the regular expression before the ^ $ (start-end) designations were included to limit it to one character.

Reference:
Special NATO Flags: http://www.marinewaypoints.com/learn/flags/flags.shtml