p5js/nautical-flags - Add rendering of EM flag
Adds:

CTRL + EM
Emergency - Action or information that requires immediate attention.

This is a white and red checkered triangular flag.

Coding Observation:
The maxY and minY values within the 'drawBarInTriangle(...)' are absolute positions ... but may be easier to work with in terms of relative proportions. TBD, not making any changes yet.