p5js/nautical-flags-02 - Add 'CommandKey' concept to support ENTER
This adds a new category of methods that the TypeSet class can support, and starts it off with one entry:

CMD_KEY_ENTER ... which is effectively a newline character.

Wondering if this should be a special character class, like escaped characters: \n and \t (TAB).

Will see how this plays out.