p5js/nautical-flags-02 - Show keyboard on mobile device
Uses an HTML <input> field to display the keyboard on a mobile device.

When the screen is touched ...
The input is given focus ...
thus showing the mobile keyboard.

Since the <input> is off screen (based on its position), and p5.js is listening for input, it passes those key events off to the sketch.