p5js/nautical-flags-02 - Prevent touch dragging from triggering the keyboard
(Working towards allowing drag to scvroll through the message on touch-devices).

Previously, if you touched the screen (on touch screens), the keyboard would always show.

Now it tracks to see if there was a dragging action, 'touchMove', between when there was a touchStarted and touchEnded event.

CAVEAT:
This doesn't keep track of which touch started or ended, or was moved.