p5js/nautical-flags-02 - Add iNoBounce to prevent triggering refresh
This prevents the mobile browser from showing the Reload UI when the page is dragged downward.

(Step towards allowing for vertical scroll via touch interface).

NOTE:
Their docs call for adding the following to an element, but I didn't see much
difference in behavior, with and without it. (but need to revisit, because overall
the Touch event handling isn't working on iOS in the sketch).

  height: 115px;
  overflow: auto;
  -webkit-overflow-scrolling: touch;