p5js/snowflake-pen - Allow the user to start drawing anywhere
This update allows the user to start drawing anywhere on the canvas.

Previously, there was an invisible slice starting at "12-o-clock" (thinking of a circle centered in the canvas). And only if the mouse started in that area would it be considered pressed.

This fix removes that restriction. The 'drawRepeatedly(...)' function already just started where the mouse was and rotated around n-number of times.