p5js/logic-gates - Add new PushSwitch component
Introduce a new example setup within the system to test out this new component.

This 'Push Switch' component is sort of like a Button, but mean to be placed in-line within a circuit to allow the user to press a button to allow the logic to continue, or disconnect the wire.

There are two options:
- By default it closes the circuit, and pressing it opens it (stopping flow)
- By default it is open, and pressing it closes the circuit (allowing the flow to continue).

https://en.wikipedia.org/wiki/Push_switch