p5js/logic-gates - Add rendering of Push Swtich internals
Illustrates the internals of the PushSwitch, where pushing down either closes or opens the circuit.

Reference:
* "Push Switch"  https://spemco.com/circuit-symbols-101/
* https://en.wikipedia.org/wiki/Push_switch

I chose not to draw additional connection points (circles) or move the existing ones, to avoid visual complexity.

This also establishes a framework by which different types of components can have additional rendering functions. This is in lieu of subclassing and extending the render() method.