p5js/common - Add setSize() to Rectangle / Circle
To make them act more like "ducks" that can have a generic "size".

For Rectangles, this effectively makes them a square of that size;

For Circles, it is treated as the diameter, and thus the radius is half the incoming size.

(Additionally, Rectangles have to do a bit more housekeeping to maintain their location).

Re: "Ducks", see:
https://en.wikipedia.org/wiki/Duck_typing