p5js/logic-gates - Adjust sizing to improve mobile experience
This adjusts the sizing / spacing of the components based on the dimensions of the canvas.

The rectangle-shaped components need to have their size passed in so that they remain in the correct position after resizing.

This is because the x,y positions of Rectangles are based on the Top Left corner; and so setting a Rectangle to position 10, 10 (calculated based on size of 40px) and if they internally (which they did) set their size 50px, when they are resized to 40px, they will be centered incorrectly.