paperjs/rack-track - Refactor tool to class
This encapsulates the drawing methods into a new 'TrackPen', and so will allow for other tools to be added as the sketch progresses in complexity.

Note:
At first I tried to "extend paper.Tool" ... but turns out that paper.Tool is not a class, but a function that returns an object.

I preferring to create ES6 classes here, so instead of using the PaperScopeItem.extend() capability ot paper.js, I'm going with encapsulation (and binding) to achieve this.

Ref:
https://github.com/paperjs/paper.js/blob/aa9dc86e7bc210bb7e4e0215bd34240e93636462/src/tool/Tool.js#L38