paperjs/race-track - Refine drawing after lifting mouse
Adds a callback for the mouseUp event to smooth out the line, and close the loop.

Also uses the minDistance config option, and sets it to 40 points, to avoid very jagged lines:
http://paperjs.org/reference/tool/#mindistance

Lastly, adds some key controls to perform addtional modification of the last draw path , to see what simplify, smooth, and flatten do.

Observation:
- simplify() is destructive, which is understood by its function call.
- smooth() seems to add complexity, but only upto a certain point.
- flatten() seems to have little effect after repeat calls.