paperjs/race-track - Establish a 'train engine'
The 'train' is visualized by a simple red circle, moving around the path drawn by the user.

After drawing a loop, pressing 't' will start the train, and pressing 'Escape' will pause the animation.

For now, the train moves with a constant velocity regardless of the curvature of the path.