paperjs/common - Add basic Particle class
This is not in use yet (and normally I don't like committing unused code), but think this is sufficiently small, yet good representation to capture the idea of the a particle in the context of paper.js.

Further, this is very similar to the existing p5js/common/particle.js in this sketchbook, and the two could potentially be merged.

This class adds the concept of 'jerk', which is the acceleration of the acceleration.

But both models have relatively little inreations with their underly JS library (p5.js or paper.js) and those aspects can be extracted via subclassing or ducktyping.