paperjs/race-track - Add concept of tires and wear
The race car now has a set of tires, that wear out as the car goes round the track.

The tires are represented as a simple measure of their condition; and a number ranging from 1 (new) to 0 (fully degraded).

Additionally, there is a simple modeling of the weight shift around the turns. The "outside" tire (relatively to the turning direction) suffers from a 0.6 wearing factor, where as the "inside" tire suffers a bit less, 0.4, because there is less weight beared by that tire.

The paper.js getCurvatureAt(...) function is used to obtain the intensity of the curve and the direction:
http://paperjs.org/reference/path/#getcurvatureat-offset

There is no implication on the car's performance with lower grade tires.

NOTE:
- Might be a premature detail refinement to perform 4 samples of the curvature over each step.