paperjs/race-track - Refactor User input
Establishes a top-level 'UserInterface' class to manage the different (PaperJS) tools that will come up as the complexity of the UI grows.

This establishes a pattern of having an active tool, which can have specific behavior, and propagate some of the UI events up to the base UserInterface class.

I think in time, I will want to minimize the event propagation (or somehow reverse it) so that child-objects are not responsible to know if the parent class should receive the event.

For now, this as some UI improvements of:
* Preventing the drawing of two or more Tracks (and leaving older tracks unreferenced)
* Moving concepts out of TrackPen related to starting / pausing the race, or adding new cars.