paperjs/int-of-color - Add basic scaling and centering of the plates
Idea:
- preserve the aspect ratio of the plate
- center (horizontally / vertically) in this viewer's area
- display it as large as possible

Leverages PaperJS'   "Group" object, where a collection of shapes can be groupe together, and then transformed together:
http://paperjs.org/reference/group/

The _calc* functions added in this commit will likely be refactored out to a "RectangleLayoutManager" class. They are intended to be re-usable across future sketches.

The `computeBounds(...)` method was written before learning that the PaperJS Group has a couple of distinct concepts of bounds.