site - Update /sketchbook to list all sketches
Currently just assumes every sketch has a screenshot-01.png (some of which are missing) and renders a thumbnail and title of the page for the sketch.

Of note is the addition of the custom `bh_skip_page` variable to prevent the display of pages like:
- 404 error page
- RSS feed
- CSS files
- the /sketchbook landing page

This was done in favor of having pages opt-into being rendered, since that would mean updating a lot of front-matter (Jekyll term) sections of files.

Reference Links:
Front-Matter
https://jekyllrb.com/docs/front-matter/

Adding front-matter to Static files:
https://jekyllrb.com/docs/static-files/#add-front-matter-to-static-files