processing/event_sequencer - Fix issue using processing.js
Javascript does not support method overloading.

Actual but was mostly in: LineUtil.drawBezier(...), it was calling the wrong constructor method for LineView class.

If you're not familar:
"Method Overloading" is when you have more than one method with the same name, but each accepts a different set of parameters (perhaps of different types).

It is only possible in strongly typed languages, like Java. But doesn't work in weakly typed language like Javascript.

Reference:
https://en.wikipedia.org/wiki/Function_overloading
https://en.wikipedia.org/wiki/Strong_and_weak_typing