site - Add breadcrumb and title for minimal layout
When the 'layout: minimal' is declared, the page is effectively full-screen and thus should have title of the page (sketch) display.

Added a simple breadcrumb back up to the Sketchbook listing page and the overall site homepage.

Also drops the 'About Me' and now redundant 'Sketchbook' links in top right (behind the drop-down on mobile).