p5js/forest-02 - Refactor sketch margin up to Jekyll template
This makes sense to include within the Jekyll template because the margin is a byproduct of the Jekyll-powered header template.

Note:
Considered having p5.js target a specific containing <div> to insert its canvas ... as per:
https://p5js.org/examples/instance-mode-instance-container.html

But that seems have to have the less desireable effect of having to preface all p5.js calls with a global sketch variable.

Other consideration:
I want the JS and HTML files to be as unaware of my Jekyll site as possible so as to allow anyone to just download the files and view them without having to run the Jekyll site, and to avoid having to check for particular <div> heights.

...

This story is likely not finished. I might consider having p5.js add the canvas, and then get the DOM canvas element and see where it sits (in terms of y-position) on the page, and then resize itself, all within the setup().