p5js/forest-02 - Fix browser content / encoding warning
Only impacts local-dev.html files...

Avoid console warning like:
"The character encoding of the HTML document was not declared. The document will render with garbled text in some browser configurations if the document contains characters from outside the US-ASCII range. The character encoding of the page must be declared in the document or in the transfer protocol."