p5js/common - Add Bezier approximation of Circle
Based on: https://spencermortensen.com/articles/bezier-circle/

Also fixes a bug related to moving Polybeziers ... because two BezierCurves would end up referencing the same Point and both would move it. Now, the `points` array of the BezierCurve is a list of the Points that that object manages.  When a curve is attached to another, it no longer manages the attachment point (and thus is no responsible for moving it, drawing it, or allowing it to be dragged).