p5js/common - Paisley refactor/improve handling of point dragging
Add supoprt for knowing which point is being dragged, and simplify the propagation of mouse events to only that item, as well as enable item-specific behavior.

This allows us to only calculate the internal helper points when the position of the Paisley (bulb) is updated, and avoid that unnecessary calcuation when moving the tail.