p5js/common - MINOR safe-guard against undefined method calls
Generally, I think that the shapes should only receive mouse events (or any events) that they are expecting; thus the app.js is updated to filter to only call handle mouseReleased on object that are currently being dragged.

But, this also makes the Paisley class fault tolerant of unexpected receiving that method call.