p5js/common - Allow the Paisley heading to be dragged around
This also establishes a dynamic behavior of the paisley.heading = x assignment, so that hte Paisly can recalculate the internal points when the heading changes.

Note: this still falls victim of not addjusting the bezuier curves at the tail when the heading is signfiicantly changed, resulting in overlapping lines right now.