mapbox/layers - Fix issue with layer buttons
Noticed that the buttons no longer worked to toggle the layers, and traced this back to the case-sensitivity of the layer ID (and passing in a UPPER case version of the layer name).

This fixes the issue by pushing the layer ID to a data attribute and fetching that on press.