p5js/slippy-map - Correct calculation of (scaled) tiles that can fit
Adjust the calculation of how many tiles fit across and up/down in viewable area, based on the zoom level.

Recompute this when ever the zoom level is reset.

(This fixes the bug where more tiles than necessary were attempted to be rendered, and thus removes the need for the safeguard that the tile width/height wasn't negative).