p5js/slippy-map - Fix drift during dragging
This recalculates the drag offset compared to the start of each drag interaction.

Otherwise, it appears that because the mouse positions passed in by p5.js are whole numbers, there can be compounding errors due to rounding/flooring of the mouse position differes ...

Eg, move a mouse over four frames, may move it 0.5 pixel each time, resulting in ~ 2 pixels
But with the mouseX and mouseY, this may be interpreted as 3-4 pixels.

Didn't test this at a fine grain, but moving it across a 460 pixel width rectangle, the previous implementation interpreted as ~530 pixels of movement.