p5js/../bezier - Fix bug preventing dragging on mobile device
What did work
- Enabling nobounce (along with viewport) to avoid typical iOS reload on drag down
- propagate touch events onto the default mouse handlers

---
Despite docs stating that mousePressed() and mouseReleased() would be called on touch interactions if the touch...() functions are not defined; that didn't seem to be the case. (touchMoved() does appear to call mouseDragged())

Additionally, tried to update the p5.js version, but that didn't resolve the issue.