p5js/common/voronoi-3 - Optimize by only drawing cells if needed
This adds a mode to too only draw if the cell needs to be draw

And allows external context to set that

For example, here, when after drawing highlighted cells, we mark them for redraw in the next frame. They may still be highlighted, but will be redrawn anyway (this is a slight inefficiency, but overall this speeds up the draw cycle considerably).