p5js/common/voronoi-4 Adjust to no-regen sites on d3-Delaunay toggle
Avoid regenerating the sites when toggling between d3-delaunay and default (gorhill implementation)

This is to have an apples-to-apples comparison in performance measurements between the two implementations.

Also, disable optimization in gorhill-drawVoronoi() which only re-drew
cells marked as needed to be redrawn. This was necessary because of the
beginShape() and endShape() of the d3-p5js render voronoi not resulting
in the cells being filled in and thus toggling between the two was
problematic. [Hindsight, might have been better to do a canvas redraw
once in handling the dat-gui drawing mode toggle; alas]