MINOR - Add excerpts to posts to prevent layout bugs.
Without these, the template would break, as Jekyll would try to auto-generate an excerpt from the content, and given the nature of the JS / Images, this would break the layout.

Ramification of this was a simple line of text preceding the header, not a big deal, but not great. Interestingly, my local jekyll would break-more, the index page would also break the excerpts whereas Github pages didn't seem to fall victim to the bug in that context

Alternate approach to consider in the future is to switch to:
{{ post.content | strip_html | truncatewords: 50 }}

In place of {{ post.excerpt }}
From: https://stackoverflow.com/a/23504519

(Though that presumes the first 50 words are a good summation of the post).