p5js/cc/marching-squares-2 - MINOR optimization using lookup array
Instead of having switch case and its code branching, using a lookup table (as done with the simple, non-interpolated drawing).

Using the chrome perf tools, this does seem to alter the runtime of this part of the code;
Anecdotally, maybe its 5% faster
... but overall this function used to take ~ 1.0ms to 1.5 ms
... and spot checking, maybe runs in 0.9ms - to 1.4ms

Challenge is that the duration is dependent on how many lines need to be drawn.