p5js/common/examples/paisley-2 Add exterior accent (work-in-progress)
I'm thinking that this might be the way to go here ... I'm seeing issues with getting the tangent and moving a 'margin' away from the edge.

I'm going to investigate creating a scaled 'Paisley' object that sits below this Paisley Object.

I am interesting in generizing the the ability to have accent-marks (interior and exterior) around my p5.js shapes ...
... but maybe all of the shapes follow a model of having an exterior/interior representation of themselves.

Or they just scale themselves in place as they draw their accents ... hmmm.