p5js/ecosystem-2 Create a second iteration - to
Mostly splitting this out because I want to leave the prior sketch as a snapshot of the progress at the time (a few years ago now) ... keeping in mind that this repo is a 'sketchbook', there's a degree of nostaglia.

Spliting this out wiill allow for:
- refactoring over to ColorRamp for color gradient
- adding in resources
- experimenting with OpenSimplex Noise generation

Ultimately I imagine a v3 where the cells are not an array of objects, but rather uses the 'DiscreteField' object model of the Marching Squares sketches which is an object that has a bunch of arrays that each represent a value at a position in the field.

(opening the door to mixed data manipulation in the GPU alongside regular JS)