js/models/space-colonization - Remove concept of 'RootTip'
This was a very rudimentary implementation to begin with, and could be added back in through extension 😏 by sketches that use this common lib.

I don't think it belongs in the base implementation; may sort of implement it by allowing leaf-segments to extend a bit, or branch (if there are more resources to be acquired).