p5js/forest-fires-2 - Adjust fire threshold of dry grass
Making this the same as GRASS_WET because with the introduction of moisture level (and different initial moisture levels) we've drawn a distinction between these two fuels. And once there is a TOOL_PLANE_WATER_DROP which adds moisture, I want these two to behave the (virtually ) same.

COUNTERPOINT: There is probably a difference in dry grass that has water on it, compared to grass which has internal moisture; maybe in the rate at which it loses moisture due to heat.