p5js/models/voronoi - Fix bug in clipping to Polygon; edges share vertex objs
Prior to this fix, we were modifying the Edge vertexes in the first pass
This resulted in some line segments that would be fully clipped out (because they are outside of the polygon) to be pulled back in, overlap the polygon and be clipped; resulting in odd arrangement of lines.

Now we mark edges to be clipped, caching a the line segment to clip them to.
Then remove all fully clipped objects
And then finally modify the edges

Still need to implement the 'closeCellsToPolygon()'