js/models/voronoi_ext - Fix bug in clipping to Polygon with points on bounds
Previously, was seeing issues related to edges connected to points on the polygon that are on its bounding box.

This fixes the issue by creating a slightly larger bounding box
So that all edges are clearly outside of the bounds, and can be clipped without dealing with too many thresholds for inside/outside of the polygon for points that are basically on the edge.

May need to tackle the threshold checks eventually, but this works.